scripts/requirements.txt
textblob==0.17.1
fpdf==1.7.2
google-generativeai==0.3.1
python-dotenv==1.0.0
